requirements.txt
google-genai
groq
feedparser
yt-dlp
requests
python-dotenv
Pillow
ffmpeg-python